feat(punctuation): 支持输入右单引号和右双引号
