+++ Fire/types.swift
@@ -189,28 +189,28 @@ let punctuation: [String: String] = [
     "/": "、",
     ";": "；",
     "'": "‘",
-    "[": "［",
-    "]": "］",
-    "`": "｀",
+    "[": "【",
+    "]": "】",
+    "`": "·",
     "!": "！",
-    "@": "‧",
-    "#": "＃",
+    "@": "@",
+    "#": "#",
     "$": "￥",
-    "%": "％",
+    "%": "%",
     "^": "……",
-    "&": "＆",
-    "*": "×",
+    "&": "&",
+    "*": "*",
     "(": "（",
     ")": "）",
-    "-": "－",
+    "-": "-",
     "_": "——",
-    "+": "＋",
-    "=": "＝",
-    "~": "～",
-    "{": "｛",
+    "+": "+",
+    "=": "=",
+    "~": "~",
+    "{": "「",
     "\\": "、",
-    "|": "｜",
-    "}": "｝",
+    "|": "|",
+    "}": "」",
     ":": "：",
     "\"": "“",
     "<": "《",
